backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
langgraph>=0.4.0
langchain-google-genai>=2.1.0
pydantic-ai[google]>=0.1.0
pydantic>=2.10.0
google-genai>=1.0.0
e2b-code-interpreter>=1.2.0
pygithub>=2.5.0
gitpython>=3.1.0
tree-sitter>=0.24.0
tree-sitter-python>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0
structlog>=24.4.0
sse-starlette>=2.2.0
python-dotenv>=1.0.0
slowapi>=0.1.9
httpx>=0.28.0
